I noticed that the code that handles ROM traps for the serial bus handles the following 2 BASIC instructions differently:
    OPEN 6,6

    OPEN 6,6,0

In the first case, the (default) secondary address is never sent.

This is because the trap code collects the device address, then later
when it gets the secondary address, in the next trap, it uses them both
in a single call to send them, either via opencbm, or to its further
internal emulation.

When the SA is defaulted, the second trap never comes.

On a real bus, the LISTEN or TALK would of course already have been
sent, but in the emulation it gets lost.

This change aims to fix that by detecting in the traps for read/write
that the secondary address was never sent, and then do it there (SA = 0).
This causes an SA to be sent that wouldn't be sent on a real machine but
that is better than missing the LISTEN or TALK.


git-svn-id: svn+ssh://svn.code.sf.net/p/vice-emu/code/trunk@28037 379a1393-f5fb-40a0-bcee-ef074d9b53f7