Create window with the correct size, right first time (xaw).
Statically creating a 300x100 window and then resizing it is not just a
cosmetic problem on startup. It can confuse window managers with "inteligent"
window placement strategy to put our window somewhere where it'll not fit
after resize. For example in the lower right corner, which is rather annoying
as most of the window ends up off screen.


git-svn-id: svn+ssh://svn.code.sf.net/p/vice-emu/code/trunk@28628 379a1393-f5fb-40a0-bcee-ef074d9b53f7