Clean up the R65C02 drive CPU mess
There was severe confusion 6510 and R65C02 drive CPUs. It only worked by luck,
and because of the incorrectly shared drive registers. That was the reason why
the monitor didn't show any cpu state for the R65C02. Also for the R65C02
drives sometimes the 6510 was executing the code. Luckily the DOS does not have
R65C02 opcodes...


git-svn-id: svn+ssh://svn.code.sf.net/p/vice-emu/code/trunk@28945 379a1393-f5fb-40a0-bcee-ef074d9b53f7