Gtk3: Add the final checkboxes to the drive settings UI
Add checkboxes for the IECDevice[8-11], AttachDeive[8-11]Readonly and Drive[8-11]RTCSave resources. Only the C64(sc/cpu) and C128 are supported, since I now have to create separate layouts depending on machine type. The logic to keep widgets synchronized is also gone due to the new layouts. That's going to need rewriting as well.



git-svn-id: svn+ssh://svn.code.sf.net/p/vice-emu/code/trunk@33972 379a1393-f5fb-40a0-bcee-ef074d9b53f7