Update rawnet.c: make rawnet_get_standard_inferface() return a heap-allocated interface name.
Since pcap_lookupdev() is deprecated (it returns a pointer to a static buffer), I'v altered the rawnet modules to return a heap-allocated copy of the default interface name, in the non-Windows code. The Windows code needs updating by someone who is actually comfortable with WinPCap.
It also needs a bit of testing, works on *nix, no idea about DOS/Windows.



git-svn-id: svn+ssh://svn.code.sf.net/p/vice-emu/code/trunk@34384 379a1393-f5fb-40a0-bcee-ef074d9b53f7