GTK3: Quick hack to add CRT control toggle to status bar.
Quick hack of compyx's own quick hack; there's still a lot of refinement to
do here. The button's color doesn't match the other widgets, and there are
a large number of magic numbers within the drive layout code in the status
bar that make adding new controls like this more complicated than it should
be.


git-svn-id: svn+ssh://svn.code.sf.net/p/vice-emu/code/trunk@34952 379a1393-f5fb-40a0-bcee-ef074d9b53f7