Gtk3: Make a setting tree node selectable via a "path"
Add the possibility to select a settings tree node by path (for example
"input/joystick"), both when popping up the settings UI and when it is already
active. This allows use to jump straight to a certain settings dialog/page
from anuwhere in the UI.
For now, the joystick popup menu has "Configure joysticks" as an example
implementation. Obivously this needs testing and checking the many settings
UI tree nodes for proper ID's.




git-svn-id: svn+ssh://svn.code.sf.net/p/vice-emu/code/trunk@35331 379a1393-f5fb-40a0-bcee-ef074d9b53f7