Gtk3: Add recording status widget to the statusbar
This widget displays what is recording and if it is recording or stopped. It contains a button to stop recording (which currently stops event, audio and video recording). Somehow calling event_record_active() does not work for displaying events are being recorded, but the STOP button works for event recording.



git-svn-id: svn+ssh://svn.code.sf.net/p/vice-emu/code/trunk@36070 379a1393-f5fb-40a0-bcee-ef074d9b53f7