Should fix bug #228.
There was some code in vdrive.c that did not believe in the  existence
of a drive #1.
Also, added drive number to verbose logging of attaches/detaches, and
more debug logging in fdc.c

The problem was that some data for drive 1 was not initialized, in particular
BAM info. Therefore, storing the new disk ID for a formatted disk failed.
Then, due to ID mismatches, writing the BAM and directory sectors failed too.


git-svn-id: svn+ssh://svn.code.sf.net/p/vice-emu/code/trunk@37834 379a1393-f5fb-40a0-bcee-ef074d9b53f7